requirements.txt
pybotters>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
loguru>=0.7.0